Move `task` to `rt`
