+++ core/src/pool.rs
@@ -1,7 +1,6 @@
 use crate::runtime::{Extra, Runtime};
-use crate::task::Executor;
 use crate::Result;
-use abel_rt::mlua;
+use abel_rt::{mlua, Executor};
 use futures::{Future, FutureExt};
 use log::error;
 use std::any::Any;
