+++ core/src/runtime.rs
@@ -4,7 +4,8 @@ use crate::ErrorKind::{self, *};
 use crate::{AbelState, Error, Result};
 use abel_rt::lua::error::{resolve_callback_error, rt_error_fmt};
 use abel_rt::lua::http::{LuaRequest, LuaResponse};
-use abel_rt::lua::{Isolate, LuaTableExt, Sandbox};
+use abel_rt::{Isolate, Sandbox};
+use abel_rt::lua::LuaTableExt;
 use abel_rt::mlua::{self, ExternalError, FromLuaMulti, Function, Table, TableExt, ToLuaMulti};
 use abel_rt::{Cleanup, CustomError, Source};
 use clru::CLruCache;
