+++ rt/src/lib.rs
@@ -2,10 +2,12 @@ pub mod lua;
 
 mod path;
 mod source;
+mod task;
 
 pub use lua::error::CustomError;
 pub use lua::http::{LuaRequest, LuaResponse};
-pub use lua::{Cleanup, Isolate, Sandbox};
+pub use lua::sandbox::{Cleanup, Sandbox};
+pub use lua::isolate::{Isolate, IsolateBuilder};
 pub use mlua;
 pub use source::{Source, SourceVfs};
-// pub use lua::context;
+pub use task::Executor;
