+++ rt/src/task/executor.rs
@@ -1,6 +1,6 @@
 use super::task_future::TaskFuture;
 use super::Task;
-use abel_rt::{mlua, Cleanup, Sandbox};
+use crate::{Cleanup, Sandbox};
 use futures::future::select;
 use futures::future::Either::*;
 use futures::stream::FuturesUnordered;
